+++ benchmark/src/main/java/com/hubspot/jinjava/benchmarks/liquid/Tags.java
@@ -5,7 +5,6 @@ import com.hubspot.jinjava.lib.tag.SetTag;
 import com.hubspot.jinjava.lib.tag.Tag;
 import com.hubspot.jinjava.tree.TagNode;
 
-
 /**
  * Liquid::Template.register_tag 'paginate', Paginate
  * Liquid::Template.register_tag 'form', CommentForm
@@ -16,6 +15,8 @@ import com.hubspot.jinjava.tree.TagNode;
 public class Tags {
 
   public static class PaginateTag implements Tag {
+    private static final long serialVersionUID = -4143036883302838710L;
+
     @Override
     public String interpret(TagNode tagNode, JinjavaInterpreter interpreter) {
       return null;
@@ -33,6 +34,8 @@ public class Tags {
   }
 
   public static class CommentFormTag implements Tag {
+    private static final long serialVersionUID = 4740110980519195813L;
+
     @Override
     public String interpret(TagNode tagNode, JinjavaInterpreter interpreter) {
 
@@ -51,6 +54,8 @@ public class Tags {
   }
 
   public static class AssignTag extends SetTag {
+    private static final long serialVersionUID = -8045822376271136191L;
+
     @Override
     public String getName() {
       return "assign";
@@ -58,15 +63,19 @@ public class Tags {
   }
 
   public static class TableRowTag implements Tag {
+    private static final long serialVersionUID = 7058892410901688159L;
+
     @Override
     public String interpret(TagNode tagNode, JinjavaInterpreter interpreter) {
       // TODO Auto-generated method stub
       return null;
     }
+
     @Override
     public String getName() {
       return "tablerow";
     }
+
     @Override
     public String getEndTagName() {
       return "endtablerow";