+++ src/main/java/com/hubspot/jinjava/el/JinjavaELContext.java
@@ -3,7 +3,6 @@ package com.hubspot.jinjava.el;
 import java.lang.reflect.Method;
 
 import javax.el.ELResolver;
-import javax.el.FunctionMapper;
 
 import de.odysseus.el.util.SimpleContext;
 
@@ -20,7 +19,7 @@ public class JinjavaELContext extends SimpleContext {
   }
 
   @Override
-  public FunctionMapper getFunctionMapper() {
+  public MacroFunctionMapper getFunctionMapper() {
     if (functionMapper == null) {
       functionMapper = new MacroFunctionMapper();
     }
@@ -29,10 +28,7 @@ public class JinjavaELContext extends SimpleContext {
 
   @Override
   public void setFunction(String prefix, String localName, Method method) {
-    if (functionMapper == null) {
-      functionMapper = new MacroFunctionMapper();
-    }
-    functionMapper.setFunction(prefix, localName, method);
+    getFunctionMapper().setFunction(prefix, localName, method);
   }
 
 }