+++ src/main/java/com/hubspot/jinjava/el/JinjavaInterpreterResolver.java
@@ -13,32 +13,50 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 
+import javax.el.ArrayELResolver;
+import javax.el.CompositeELResolver;
 import javax.el.ELContext;
+import javax.el.ELResolver;
+import javax.el.MapELResolver;
+import javax.el.PropertyNotFoundException;
+import javax.el.ResourceBundleELResolver;
 
 import org.apache.commons.lang3.LocaleUtils;
 import org.apache.commons.lang3.StringUtils;
 
-import com.google.common.collect.Lists;
 import com.hubspot.jinjava.el.ext.AbstractCallableMethod;
 import com.hubspot.jinjava.el.ext.ExtendedParser;
+import com.hubspot.jinjava.el.ext.JinjavaBeanELResolver;
+import com.hubspot.jinjava.el.ext.JinjavaListELResolver;
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
 import com.hubspot.jinjava.interpret.TemplateError;
+import com.hubspot.jinjava.interpret.TemplateError.ErrorReason;
+import com.hubspot.jinjava.interpret.TemplateError.ErrorType;
 import com.hubspot.jinjava.objects.PyWrapper;
 import com.hubspot.jinjava.objects.collections.PyList;
 import com.hubspot.jinjava.objects.collections.PyMap;
 import com.hubspot.jinjava.objects.date.FormattedDate;
 import com.hubspot.jinjava.objects.date.PyishDate;
 import com.hubspot.jinjava.objects.date.StrftimeFormatter;
-import com.hubspot.jinjava.util.JinjavaPropertyNotResolvedException;
-import com.hubspot.jinjava.util.VariableChain;
 
 import de.odysseus.el.util.SimpleResolver;
 
 public class JinjavaInterpreterResolver extends SimpleResolver {
 
+  private static final ELResolver DEFAULT_RESOLVER_READ_WRITE = new CompositeELResolver() {
+    {
+      add(new ArrayELResolver(false));
+      add(new JinjavaListELResolver(false));
+      add(new MapELResolver(false));
+      add(new ResourceBundleELResolver());
+      add(new JinjavaBeanELResolver(false));
+    }
+  };
+
   private JinjavaInterpreter interpreter;
 
   public JinjavaInterpreterResolver(JinjavaInterpreter interpreter) {
+    super(DEFAULT_RESOLVER_READ_WRITE);
     this.interpreter = interpreter;
   }
 
@@ -46,55 +64,62 @@ public class JinjavaInterpreterResolver extends SimpleResolver {
   public Object invoke(ELContext context, Object base, Object method,
       Class<?>[] paramTypes, Object[] params) {
 
-    Object methodProperty = getValue(context, base, method, false);
-    if (methodProperty != null && methodProperty instanceof AbstractCallableMethod) {
-      context.setPropertyResolved(true);
-      return ((AbstractCallableMethod) methodProperty).evaluate(params);
+    try {
+      Object methodProperty = getValue(context, base, method, false);
+      if (methodProperty != null && methodProperty instanceof AbstractCallableMethod) {
+        context.setPropertyResolved(true);
+        return ((AbstractCallableMethod) methodProperty).evaluate(params);
+      }
+    } catch (IllegalArgumentException e) {
+      // failed to access property, continue with method calls
     }
 
     // TODO map named params to special arg in fn to invoke
     return super.invoke(context, base, method, paramTypes, params);
   }
 
+  /**
+   * {@inheritDoc}
+   *
+   * If the base object is null, the property will be looked up in the context.
+   */
   @Override
-  public Object getValue(ELContext context, Object base, Object prop) {
-    return getValue(context, base, prop, true);
+  public Object getValue(ELContext context, Object base, Object property) {
+    return getValue(context, base, property, true);
   }
 
-  private Object getValue(ELContext context, Object base, Object prop, boolean errOnUnknownProp) {
-    String property = Objects.toString(prop, "");
+  private Object getValue(ELContext context, Object base, Object property, boolean errOnUnknownProp) {
+    String propertyName = Objects.toString(property, "");
     Object value = null;
 
-    if (ExtendedParser.INTERPRETER.equals(prop)) {
+    if (ExtendedParser.INTERPRETER.equals(property)) {
       value = interpreter;
-    }
-    else if (property.startsWith(ExtendedParser.FILTER_PREFIX)) {
-      value = interpreter.getContext().getFilter(StringUtils.substringAfter(property, ExtendedParser.FILTER_PREFIX));
-    }
-    else if (property.startsWith(ExtendedParser.EXPTEST_PREFIX)) {
-      value = interpreter.getContext().getExpTest(StringUtils.substringAfter(property, ExtendedParser.EXPTEST_PREFIX));
-    }
-    else {
+    } else if (propertyName.startsWith(ExtendedParser.FILTER_PREFIX)) {
+      value = interpreter.getContext().getFilter(StringUtils.substringAfter(propertyName, ExtendedParser.FILTER_PREFIX));
+    } else if (propertyName.startsWith(ExtendedParser.EXPTEST_PREFIX)) {
+      value = interpreter.getContext().getExpTest(StringUtils.substringAfter(propertyName, ExtendedParser.EXPTEST_PREFIX));
+    } else {
       if (base == null) {
-        value = interpreter.retraceVariable((String) prop, interpreter.getLineNumber());
-      }
-      else {
+        // Look up property in context.
+        value = interpreter.retraceVariable((String) property, interpreter.getLineNumber());
+      } else {
+        // Get property of base object.
         try {
-          value = new VariableChain(Lists.newArrayList(property), base).resolve();
-        } catch (JinjavaPropertyNotResolvedException e) {
+          value = super.getValue(context, base, propertyName);
+        } catch (PropertyNotFoundException e) {
           if (errOnUnknownProp) {
-            interpreter.addError(TemplateError.fromUnknownProperty(base, property, interpreter.getLineNumber()));
+            interpreter.addError(TemplateError.fromUnknownProperty(base, propertyName, interpreter.getLineNumber()));
           }
         }
       }
     }
 
     context.setPropertyResolved(true);
-    return wrap(interpreter, value);
+    return wrap(value);
   }
 
   @SuppressWarnings("unchecked")
-  public static Object wrap(JinjavaInterpreter interpreter, Object value) {
+  Object wrap(Object value) {
     if (value == null) {
       return null;
     }
@@ -139,7 +164,7 @@ public class JinjavaInterpreterResolver extends SimpleResolver {
       try {
         return StrftimeFormatter.formatter(d.getFormat());
       } catch (IllegalArgumentException e) {
-        interpreter.addError(TemplateError.fromException(e));
+        interpreter.addError(new TemplateError(ErrorType.WARNING, ErrorReason.SYNTAX_ERROR, e.getMessage(), null, interpreter.getLineNumber(), null));
       }
     }
 
@@ -151,7 +176,7 @@ public class JinjavaInterpreterResolver extends SimpleResolver {
       try {
         return LocaleUtils.toLocale(d.getLanguage());
       } catch (IllegalArgumentException e) {
-        interpreter.addError(TemplateError.fromException(e));
+        interpreter.addError(new TemplateError(ErrorType.WARNING, ErrorReason.SYNTAX_ERROR, e.getMessage(), null, interpreter.getLineNumber(), null));
       }
     }