+++ src/main/java/com/hubspot/jinjava/lib/filter/JoinFilter.java
@@ -5,14 +5,12 @@ import java.util.List;
 import java.util.Objects;
 
 import com.google.common.base.Joiner;
-import com.google.common.collect.Lists;
 import com.hubspot.jinjava.doc.annotations.JinjavaDoc;
 import com.hubspot.jinjava.doc.annotations.JinjavaParam;
 import com.hubspot.jinjava.doc.annotations.JinjavaSnippet;
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
 import com.hubspot.jinjava.util.ForLoop;
 import com.hubspot.jinjava.util.ObjectIterator;
-import com.hubspot.jinjava.util.VariableChain;
 
 @JinjavaDoc(
     value = "Return a string which is the concatenation of the strings in the sequence.",
@@ -58,7 +56,7 @@ public class JoinFilter implements Filter {
       Object val = loop.next();
 
       if (attr != null) {
-        val = new VariableChain(Lists.newArrayList(attr), val).resolve();
+        val = interpreter.resolveProperty(val, attr);
       }
 
       vals.add(Objects.toString(val, ""));